+++ src/main/java/com/armaandhir/demo/service/UserAccountService.java
@@ -1,6 +1,5 @@
 package com.armaandhir.demo.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.armaandhir.demo.model.UserAccount;
@@ -9,11 +8,11 @@ public interface UserAccountService {
 
 	Collection<UserAccount> findAll();
 	
-	UserAccount findOne(BigInteger id);
+	UserAccount findOne(Long id);
 	
 	UserAccount create(UserAccount account);
 	
 	UserAccount update (UserAccount account);
 	
-	void delete (BigInteger id);
+	void delete (Long id);
 }
